Introduce ElasticsearchJavaBasePlugin
just configuring common conventions on java based projects without
adding opinionated sourcesets. Reduces the configuration overhead for
test only projects.
