+++ build-tools-internal/src/main/java/org/elasticsearch/gradle/internal/test/StandaloneTestPlugin.java
@@ -30,7 +30,6 @@ public class StandaloneTestPlugin implements Plugin<Project> {
             test.mustRunAfter(project.getTasks().getByName("precommit"));
         });
 
-        ElasticsearchJavaPlugin.configureCompile(project);
         project.getTasks().named("check").configure(task -> task.dependsOn(project.getTasks().named("test")));
     }
 }
